+++ src/main/java/com/mikeoertli/rangeselector/Constants.java
@@ -7,9 +7,9 @@ package com.mikeoertli.rangeselector;
  */
 public class Constants
 {
-    public static final String RANGE_SELECTOR_ICON_32X32_PATH = "icon/range-selection-icon-32.png";
-    public static final String RANGE_SELECTOR_ICON_64X64_PATH = "icon/range-selection-icon-64.png";
-    public static final String RANGE_SELECTOR_ICON_128X128_PATH = "icon/range-selection-icon-128.png";
+    public static final String RANGE_SELECTOR_ICON_32X32_PATH = "icon/logo/range-selection-icon-32.png";
+    public static final String RANGE_SELECTOR_ICON_64X64_PATH = "icon/logo/range-selection-icon-64.png";
+    public static final String RANGE_SELECTOR_ICON_128X128_PATH = "icon/logo/range-selection-icon-128.png";
     public static final String APP_NAME = "Range Selector";
     public static final String DEMO_APP_NAME = "Range Selector Demo";
 }