+++ src/main/java/com/mikeoertli/rangeselector/ui/swing/listener/ILockListener.java
@@ -1,0 +1,11 @@
+package com.mikeoertli.rangeselector.ui.swing.listener;
+
+/**
+ * Listens to lock state changes
+ *
+ * @since 0.0.3
+ */
+public interface ILockListener
+{
+    void toggleLock();
+}