+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+* 0.0.3 - 2021-01-03
+  * Support for locking the selection range (with or without a selected range)
+  * Fixes range selection when click events are outside histogram panel
+  * Very preliminary support for dark mode
+
 * 0.0.2 - 2020-11-19
   * Basic histogram with color legend.